backend-fastapi/requirements.txt
fastapi[all]
uvicorn[standard]
motor
pydantic
python-dotenv
passlib[bcrypt]
pyjwt
stripe
requests